requirements.txt
pypdf
python-docx
faiss-cpu
numpy
fastapi
uvicorn
python-multipart
streamlit
requests
